renaming and clean up
