+++ package.json
@@ -9,7 +9,6 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "@types/howler": "^2.2.1",
     "@types/jest": "^24.9.1",
     "@types/node": "^12.12.54",
     "@types/react": "^16.9.46",
@@ -20,8 +19,6 @@
     "@types/redux": "^3.6.0",
     "@types/redux-thunk": "^2.1.0",
     "axios": "^0.19.2",
-    "firebase": "^7.19.1",
-    "howler": "^2.2.0",
     "material-ui-banner": "^1.4.3",
     "moment": "^2.27.0",
     "notistack": "^0.9.17",
