+++ src/components/Auth/signupAuth.tsx
@@ -43,43 +43,6 @@ const SingUp = (props: any) => {
   };
   const [signUpcredentials, set_signUpcredentials] = useState(initialState);
 
-  // const uploadImage = async () => {
-  //   try {
-  //     //@ts-ignore
-  //     const uploadTask = storage.ref(`images/${image.name}`).put(image);
-  //     uploadTask.on(
-  //       "state_changed",
-  //       (snapshot) => {
-  //         const progress = Math.round(
-  //           (snapshot.bytesTransferred / snapshot.totalBytes) * 100
-  //         );
-  //         setProgress(progress);
-  //       },
-  //       (error) => {
-  //         console.log(error);
-  //       },
-  //       () => {
-  //         storage
-  //           .ref("images")
-  //           //@ts-ignore
-  //           .child(image.name)
-  //           .getDownloadURL()
-  //           .then((url) => {
-  //             setUrl(url);
-  //           });
-  //       }
-  //     );
-  //   } catch (error) {
-  //     console.log(error.message);
-  //   }
-  // };
-
-  // const handleChange = (e: any) => {
-  //   if (e.target.files[0]) {
-  //     setImage(e.target.files[0]);
-  //   }
-  // };
-
   function submitForm(event: OnClick): void {
     event.preventDefault();
     dispatch(signUp(signUpcredentials));
@@ -128,17 +91,6 @@ const SingUp = (props: any) => {
         fullWidth
         required
       />
-      {/* <Button type="button" onClick={uploadImage}>
-        Upload
-      </Button>
-
-      <input type="file" onChange={handleChange} />
-      <img
-        src={url || "http://via.placeholder.com/300"}
-        alt="firebaseimage"
-        style={{ width: "300px" }}
-      /> */}
-
       <DialogActions>
         <Button variant="contained" color="primary" onClick={submitForm}>
           Submit
