--- src/components/Banner/index.css
@@ -1,47 +1,0 @@
-body {
-  font: 1em sans-serif;
-  text-align: center;
-}
-.stage {
-  border-bottom: 3px solid #444;
-  display: flex;
-  height: 330px;
-  width: 100%;
-}
-.box {
-  align-self: flex-end;
-  animation-duration: 2s;
-  animation-iteration-count: infinite;
-  background-color: #f44336;
-  height: 200px;
-  margin: 0 auto 0 auto;
-  transform-origin: bottom;
-  width: 200px;
-}
-.bounce-5 {
-  animation-name: bounce-5;
-  animation-timing-function: ease;
-}
-@keyframes bounce-5 {
-  0% {
-    transform: scale(1, 1) translateY(0);
-  }
-  10% {
-    transform: scale(1.1, 0.9) translateY(0);
-  }
-  30% {
-    transform: scale(0.9, 1.1) translateY(-100px);
-  }
-  50% {
-    transform: scale(1, 1) translateY(0);
-  }
-  57% {
-    transform: scale(1, 1) translateY(-7px);
-  }
-  64% {
-    transform: scale(1, 1) translateY(0);
-  }
-  100% {
-    transform: scale(1, 1) translateY(0);
-  }
-}
