+++ src/components/UserCard/styles.tsx
